더 깔끔하게 만듦
