+++ bookshorts/templates/book_list/summarized.html
@@ -10,4 +10,5 @@
 <div>
     {{ book.answer }}
 </div>
+</script>
 {% endblock %}
