+++ bookshorts/templates/index.html
@@ -19,7 +19,7 @@
 
         <!-- 시작 화면 (intro) -->
         <div id="intro" class="intro-container">
-            <form method="post" action="/bookinfo" onsubmit="start()">
+            <form method="post" action="./answer" onsubmit="start()">
                 <h1>bookshorts에 오신 것을 환영합니다! 요약하고 싶은 책을 적어주세요.</h1>
                 <div class="input-section">
                     <label for="title">제목</label>
