+++ bookshorts/templates/login.html
@@ -16,7 +16,7 @@
                     <button type="submit" class="button">로그인하기</button>
                 </div>
             </form>
-            <a href="{{ url_for('main.register') }}" class="login-link">회원가입하기</a>
+            <a href="{{ url_for('login.register') }}" class="login-link">회원가입하기</a>
         </div>
     </div>
     <script>
