+++ bookshorts/templates/register.html
@@ -3,7 +3,7 @@
 {% block content %}
 <div class="container">
     <div class="login-form">
-        <form action="/register" method="post">
+        <form action="{{ url_for('login.register') }}" method="post">
             <div class="input-section">
                 <label for="username">이름</label>
                 <input type="text" name="username" id="username" class="input-section input">
