+++ bookshorts/templates/summarize.html
@@ -29,7 +29,7 @@
         <div class="chat-box">
             <form method="post" action="/summarize">            
                 <div class="chat-message assistant-welcome">
-                <p class="assistant">{{ book_info }}에서의 핵심 내용을 요약하겠습니다:
+                <p class="assistant">
                     {{ content }}
                 </p>
                 </div>
